refactor: keep durable state probe with lifecycle
